feat: modify linear uplot intp scatter plot
