+++ webui/react/src/components/ScatterPlot.module.scss
