+++ webui/react/src/pages/ExperimentDetails/ExperimentVisualization/ScatterPlots.tsx
@@ -1,14 +1,16 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 
-import Chart from 'components/Chart';
 import Grid, { GridMode } from 'components/Grid';
+import ScatterPlot from 'components/ScatterPlot';
 import { ShirtSize } from 'themes';
 
 const ScatterPlots: React.FC = () => {
+  const x = useMemo(() => ([ 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 ]), []);
+  const y = useMemo(() => new Array(10).fill(null).map(() => Math.random()), []);
   return (
     <Grid gap={ShirtSize.big} mode={GridMode.AutoFill}>
       {new Array(100).fill(null).map((_, index) => (
-        <Chart key={index} />
+        <ScatterPlot key={index} values={y} x={x} y={y} />
       ))}
     </Grid>
   );
